Run nova-db-sync on each Puppet run
At the moment, nova-db-sync only runs if triggered by something else.
While this works in most situations, it means the modules aren't
idempotent. Essentially, if the db hasn't been created, we need to
create it, regardless of what else happens in the same puppet run.
Seeing as the db sync is idempotent in itself, just run it on every
puppet run.