Add prevent_db_sync flag to nova class
It turns out that SQLAlchemy migrate (the backend of "nova-manage db
sync") is racy. Running it on multiple hosts at the same time is likely
to fail. Adding a prevent_db_sync flag so that we can limit these
upgrades to happen on just a single host (e.g. "the controller")